Integrated settings menu + More
Zombie variant spawning can now be individually toggled in the new settings menu. As well as debug mode. What the data pack regards as the sea level can now be modified there as well.

Files were reorganized to make the file structure more intuitive.

Revised naming conventions. Scores of the objective zev._ are temp variables if they are prefixed with a $_ or a #_. Removed objectives that served as a function namespace in favor of a global namespace: zev._. Temp storages are denoted by having names that are strictly numbers. Temp entity tags are denoted by a name prefixed by zev. and then a number.

Load function is more efficient. All objectives except the global objective (zev._) are created on strictly the initial load.

Fixed a critical bug with Jackies where they can steal the item in your mainhand if all equipment slots of the player are populated.